[PP][OPS-9970] some model and util classes moved to cor project
