+++ app/deniedrefs/model/VerifyRefStatus.scala
@@ -18,7 +18,7 @@ package deniedrefs.model
 
 import enumeratum.{Enum, EnumEntry}
 import play.api.libs.json.Format
-import util.EnumFormat
+import tps.utils.EnumFormat
 
 import scala.collection.immutable
