+++ app/model/Journey.scala
@@ -18,6 +18,7 @@ package model
 
 import model.pcipal.{PcipalSessionId, PcipalSessionLaunchRequest, PcipalSessionLaunchResponse}
 import play.api.libs.json.{Json, OFormat}
+import tps.model.{JourneyId, Navigation}
 import tps.model.repo.HasId
 
 import java.time.Instant
