+++ app/model/StatusTypes.scala
@@ -17,8 +17,8 @@
 package model
 
 import enumeratum._
-import util.EnumFormat
 import play.api.libs.json.Format
+import tps.utils.EnumFormat
 
 import scala.collection.immutable
