+++ app/repository/JourneyRepo.scala
@@ -16,12 +16,13 @@
 
 package repository
 
-import model.SafeEquals.EqualsOps
+import tps.utils.SafeEquals.EqualsOps
 import model._
 import model.pcipal.PcipalSessionId
 import org.mongodb.scala.model.{IndexModel, IndexOptions, Indexes}
 import play.api.libs.json.Json.toJson
 import play.api.libs.json.{Format, Json, OFormat, Reads}
+import tps.model.JourneyId
 import uk.gov.hmrc.mongo.MongoComponent
 import uk.gov.hmrc.mongo.play.json.formats.MongoJavatimeFormats
